requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
--no-binary falcon

attrs==20.3.0
    # via
    #   automat
    #   service-identity
    #   twisted
automat==20.2.0
    # via twisted
beautifulsoup4==4.9.1
    # via -r requirements.in
cffi==1.14.5
    # via cryptography
click==7.1.2
    # via
    #   -r requirements.in
    #   rq
constantly==15.1.0
    # via twisted
cryptography==3.4.6
    # via
    #   pyopenssl
    #   scrapy
    #   service-identity
cssselect==1.1.0
    # via
    #   parsel
    #   scrapy
cython==0.29.21
    # via -r requirements.in
falcon-cors==1.1.7
    # via -r requirements.in
falcon==2.0.0
    # via falcon-cors
gevent==20.9.0
    # via -r requirements.in
greenlet==1.0.0
    # via gevent
gunicorn==20.0.4
    # via -r requirements.in
hiredis==1.1.0
    # via
    #   -r requirements.in
    #   redisearch
hyperlink==21.0.0
    # via twisted
idna==2.10
    # via
    #   -r requirements.in
    #   hyperlink
incremental==21.3.0
    # via twisted
itemadapter==0.2.0
    # via
    #   itemloaders
    #   scrapy
itemloaders==1.0.4
    # via scrapy
jmespath==0.10.0
    # via itemloaders
lxml==4.6.3
    # via
    #   parsel
    #   scrapy
newrelic==6.0.1.155
    # via -r requirements.in
parsel==1.6.0
    # via
    #   itemloaders
    #   scrapy
protego==0.1.16
    # via scrapy
pyasn1-modules==0.2.8
    # via service-identity
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   service-identity
pycparser==2.20
    # via cffi
pydispatcher==2.0.5
    # via scrapy
pyopenssl==20.0.1
    # via scrapy
python-dotenv==0.14.0
    # via -r requirements.in
queuelib==1.5.0
    # via scrapy
redis==3.5.3
    # via
    #   -r requirements.in
    #   redisearch
    #   rmtest
    #   rq
redisearch==2.0.0
    # via -r requirements.in
rmtest==0.7.0
    # via redisearch
rq==1.7.0
    # via -r requirements.in
scrapy==2.3.0
    # via -r requirements.in
service-identity==18.1.0
    # via scrapy
six==1.15.0
    # via
    #   automat
    #   parsel
    #   protego
    #   pyopenssl
    #   redisearch
    #   w3lib
soupsieve==2.2.1
    # via beautifulsoup4
twisted==21.2.0
    # via scrapy
w3lib==1.22.0
    # via
    #   itemloaders
    #   parsel
    #   scrapy
zope.event==4.5.0
    # via gevent
zope.interface==5.3.0
    # via
    #   gevent
    #   scrapy
    #   twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
